Change Event timestamps to be saved as 64-bit long MYA timestamps.
This caused lots of trickle down changes. I created a convenience
constructor that accepts Instants for the different Event subtypes, but
changed the fromRow factory function to use the long MYA timestamp
constructor for performance reasons.  Additionally, the compareTo
function now directly compares the long MYA timestamps and not Instants.

A handful of knock-on changes to the various services that handle
Event objects were needed.  I tried to ensure that any sections of code
that would have a performance impact would use the long MYA timestamps
as much as possible.

Note: No changes were made to the JUnit Tests to accommodate this
change.  Those updates to follow in the next commit.